Cleaned and refactored ExcelDna.COMWrappers.NativeAOT project.
