+++ Source/ExcelDna.COMWrappers.NativeAOT/ComInterfaces/DispParams.cs
@@ -1,0 +1,10 @@
+﻿namespace ExcelDna.COMWrappers.NativeAOT.ComInterfaces
+{
+    internal struct DispParams
+    {
+        public int cArgs;
+        public int cNamedArgs;
+        public int rgdispidNamedArgs;
+        public Variant[] rgvarg;
+    }
+}
