+++ Source/ExcelDna.COMWrappers.NativeAOT/ComInterfaces/Variant.cs
@@ -1,0 +1,12 @@
+﻿namespace ExcelDna.COMWrappers.NativeAOT.ComInterfaces
+{
+    internal struct Variant
+    {
+        public Variant(object? value)
+        {
+            Value = value;
+        }
+
+        public object? Value { get; set; }
+    }
+}
