+++ Source/ExcelDna.COMWrappers.NativeAOT/ComInterfaces/VariantBoolNative.cs
@@ -1,0 +1,8 @@
+﻿namespace ExcelDna.COMWrappers.NativeAOT.ComInterfaces
+{
+    internal enum VariantBoolNative : short
+    {
+        VARIANT_FALSE = 0,
+        VARIANT_TRUE = -1
+    }
+}
