+++ Source/ExcelDna.COMWrappers.NativeAOT/TypeAdapter.cs
@@ -1,4 +1,5 @@
 ﻿using ExcelDna.ComInterop;
+using ExcelDna.COMWrappers.NativeAOT.ComInterfaces;
 
 namespace ExcelDna.COMWrappers.NativeAOT
 {
@@ -6,32 +7,32 @@ namespace ExcelDna.COMWrappers.NativeAOT
     {
         public object GetProperty(string name, object comObject)
         {
-            return (comObject as ComObject)!.GetProperty(name)!;
+            return (comObject as DispatchObject)!.GetProperty(name)!;
         }
 
         public object GetIndex(int i, object comObject)
         {
-            return (comObject as ComObject)!.GetIndex(i)!;
+            return (comObject as DispatchObject)!.GetIndex(i)!;
         }
 
         public bool Is(ref Guid guid, object comObject)
         {
-            return (comObject as ComObject)!.HasInterface(ref guid);
+            return (comObject as DispatchObject)!.HasInterface(ref guid);
         }
 
         public object Invoke(string name, object[] args, object comObject)
         {
-            return (comObject as ComObject)!.Invoke(name, args)!;
+            return (comObject as DispatchObject)!.Invoke(name, args)!;
         }
 
         public void SetProperty(string name, object value, object comObject)
         {
-            (comObject as ComObject)!.SetProperty(name, value);
+            (comObject as DispatchObject)!.SetProperty(name, value);
         }
 
         public object GetObject(IntPtr pUnk)
         {
-            return new ComObject(pUnk);
+            return new DispatchObject(pUnk);
         }
 
         public void ReleaseObject(object comObject)
@@ -40,7 +41,7 @@ namespace ExcelDna.COMWrappers.NativeAOT
 
         public bool HasProperty(string name, object comObject)
         {
-            return (comObject as ComObject)!.HasProperty(name);
+            return (comObject as DispatchObject)!.HasProperty(name);
         }
     }
 }
