--- Source/ExcelDna.COMWrappers.NativeAOT/Types/Managed/DispParams.cs
@@ -1,9 +1,0 @@
-﻿namespace Addin.Types.Managed;
-
-public struct DispParams
-{
-    public int cArgs;
-    public int cNamedArgs;
-    public int rgdispidNamedArgs;
-    public Variant[] rgvarg;
-}
