--- Source/ExcelDna.COMWrappers.NativeAOT/Types/Managed/Variant.cs
@@ -1,12 +1,0 @@
-﻿namespace Addin.Types.Managed;
-
-public struct Variant
-{
-    public Variant(object? value)
-    {
-        Value = value;
-    }
-
-    public object? Value { get; set; }
-    public IntPtr DispVal { get; set; }
-}
