--- Source/ExcelDna.COMWrappers.NativeAOT/Types/Unmanaged/ComConstants.cs
@@ -1,9 +1,0 @@
-﻿namespace Addin.Types.Unmanaged;
-
-public class ComConstants
-{
-    public const int LOCALE_SYSTEM_DEFAULT = 0x0800;
-    public const int LOCALE_USER_DEFAULT = 0x0400;
-    public const int DISPID_PROPERTYPUT = -3;
-    public const uint OBJID_NATIVEOM = 0xFFFFFFF0;
-}
