--- Source/ExcelDna.COMWrappers.NativeAOT/Types/Unmanaged/DispParams.cs
@@ -1,12 +1,0 @@
-﻿using System.Runtime.InteropServices;
-
-namespace Addin.Types.Unmanaged;
-
-[StructLayout(LayoutKind.Sequential)]
-public unsafe struct DispParams
-{
-    public nint rgvarg;
-    public int* rgdispidNamedArgs;
-    public int cArgs;
-    public int cNamedArgs;
-}
