--- Source/ExcelDna.COMWrappers.NativeAOT/Types/Unmanaged/VariantBool.cs
@@ -1,7 +1,0 @@
-﻿namespace Addin.Types.Unmanaged;
-
-public enum VariantBool : short
-{
-    VARIANT_FALSE = 0,
-    VARIANT_TRUE = -1
-}
